lang: Further limit the console-only type function
This commit introduces a capsule type, `TypeType`, which is used to
extricate type information from the console-only `type` function. In
combination with the `TypeType` mark, this allows us to restrict the use
of this function to very few cases:

- Calling `type()` at the top level of a console session, to display the
  type of a given value;
- Checking if two values have equal types: `type(true) == type(false)`.

Any other use of `type()` will result in an error diagnostic.